+++ actors/init/Cargo.toml
@@ -29,9 +29,7 @@ serde = { workspace = true, features = ["derive"] }
 # V0 dependencies
 address = { package = "forest_address", version = "0.3" }
 encoding = { package = "forest_encoding", version = "0.2.1" }
-fil_types = "0.2"
 forest_cid = { version = "0.3", features = ["cbor"] }
 ipld_blockstore = "0.1"
 ipld_hamt = { version = "0.1", features = ["go-interop"] }
-runtime = { package = "forest_runtime", version = "0.2" }
 vm = { package = "forest_vm", version = "0.3" }
