+++ actors/init/src/v0/mod.rs
@@ -1,12 +1,7 @@
 // Copyright 2020 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
 
-use fil_types::{ActorID, HAMT_BIT_WIDTH};
-use forest_cid::Cid;
-use ipld_blockstore::BlockStore;
-use ipld_hamt::{BytesKey, Error as HamtError, Hamt};
 use num_derive::FromPrimitive;
-use serde::{de::DeserializeOwned, Serialize};
 use vm::METHOD_CONSTRUCTOR;
 
 pub use self::state::State;
@@ -23,23 +18,4 @@ mod types;
 pub enum Method {
     Constructor = METHOD_CONSTRUCTOR,
     Exec = 2,
-}
-
-/// Defines first available ID address after builtin actors
-pub const FIRST_NON_SINGLETON_ADDR: ActorID = 100;
-
-/// Map type to be used within actors. The underlying type is a hamt.
-pub type Map<'bs, BS, V> = Hamt<'bs, BS, V, BytesKey>;
-
-/// Create a map with a root cid.
-#[inline]
-pub fn make_map_with_root<'bs, BS, V>(
-    root: &Cid,
-    store: &'bs BS,
-) -> Result<Map<'bs, BS, V>, HamtError>
-where
-    BS: BlockStore,
-    V: DeserializeOwned + Serialize,
-{
-    Map::<_, V>::load_with_bit_width(root, store, HAMT_BIT_WIDTH)
 }
