+++ actors/init/src/v0/state.rs
@@ -1,15 +1,21 @@
 // Copyright 2020 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
 
-use super::{make_map_with_root, FIRST_NON_SINGLETON_ADDR};
 use address::{Address, Protocol};
 use encoding::tuple::*;
 use encoding::Cbor;
-use fil_types::ActorID;
 use forest_cid::Cid;
+use fvm_shared::{ActorID, HAMT_BIT_WIDTH};
 use ipld_blockstore::BlockStore;
-use ipld_hamt::Error as HamtError;
+use ipld_hamt::{BytesKey, Error as HamtError, Hamt};
+use serde::{de::DeserializeOwned, Serialize};
 use std::error::Error as StdError;
+
+/// Defines first available ID address after builtin actors
+pub const FIRST_NON_SINGLETON_ADDR: ActorID = 100;
+
+/// Map type to be used within actors. The underlying type is a hamt.
+pub type Map<'bs, BS, V> = Hamt<'bs, BS, V, BytesKey>;
 
 /// State is reponsible for creating
 #[derive(Debug, Serialize_tuple, Deserialize_tuple)]
@@ -71,3 +77,16 @@ impl State {
 }
 
 impl Cbor for State {}
+
+/// Create a map with a root cid.
+#[inline]
+pub fn make_map_with_root<'bs, BS, V>(
+    root: &Cid,
+    store: &'bs BS,
+) -> Result<Map<'bs, BS, V>, HamtError>
+where
+    BS: BlockStore,
+    V: DeserializeOwned + Serialize,
+{
+    Map::<_, V>::load_with_bit_width(root, store, HAMT_BIT_WIDTH)
+}
